+++ R/Run_Mod.R
@@ -1,64 +1,78 @@
 #' @title Run Jags Model
 #' @author Sam Schildhauer
 #' @description
-#'  run_mod() takes a data frame and adjustable mcmc inputs to
-#'  [runjags::run.jags()] as an mcmc
-#'  bayesian model to estimate antibody dynamic curve parameters.
+#'  `run_mod()` takes a data frame and adjustable MCMC inputs to
+#'  [runjags::run.jags()] as an MCMC
+#'  Bayesian model to estimate antibody dynamic curve parameters.
 #'  The [rjags::jags.model()] models seroresponse dynamics to an
 #'  infection. The antibody dynamic curve includes the following parameters:
 #'  - y0 = baseline antibody concentration
 #'  - y1 = peak antibody concentration
 #'  - t1 = time to peak
-#'  - r = shape parameter
+#'  - shape = shape parameter
 #'  - alpha = decay rate
 #' @param data A [base::data.frame()] with the following columns.
 #' @param file_mod The name of the file that contains model structure.
 #' @param nchain An [integer] between 1 and 4 that specifies
-#' the number of mcmc chains to be run per jags model.
+#' the number of MCMC chains to be run per jags model.
 #' @param nadapt An [integer] specifying the number of adaptations per chain.
 #' @param nburn An [integer] specifying the number of burn ins before sampling.
-#' @param nmc An [integer] specifying number of samples in posterior chains
-#' @param niter An [integer] specifying number of iterations.
-#' @param strat
-#' A [character] string specifying the stratification variable,
+#' @param nmc An [integer] specifying the number of samples in posterior chains.
+#' @param niter An [integer] specifying the number of iterations.
+#' @param strat A [character] string specifying the stratification variable,
 #' entered in quotes.
-#' @return
-#' - A jags.post [list()] object or multiple jags.post [list()]
-#' if stratified. Returned as a [list()] of class [runjags::runjags-class]
-#' - A [base::data.frame()] titled `curve_params` that contains the posterior
-#' distribution will be exported with the following attributes:
-#'  - `iteration` = number of sampling iterations.
-#'  - `chain` = number of mcmc chains run; between 1 and 4.
-#'  - `indexid` = "newperson", indicating posterior distribution.
-#'  - `antigen_iso` = antibody/antigen type combination being evaluated
-#'  - `alpha` = posterior estimate of decay rate
-#'  - `r` = posterior estimate of shape parameter
-#'  - `t1` = posterior estimate of time to peak
-#'  - `y0` = posterior estimate of baseline antibody concentration
-#'  - `y1` = posterior estimate of peak antibody concentration
-#'  - `stratified variable` = the variable that jags was stratified by
-#' - A [list] of `attributes` that summarize the jags inputs, including:
-#'  - `class`: Class of the output object.
-#'  - `nChain`: Number of chains run.
-#'  - `nParameters`: The amount of parameters estimated in the model.
-#'  - `nIterations`: Number of iteration specified.
-#'  - `nBurnin`: Number of burn ins.
-#'  - `nThin`: Thinning number (niter/nmc)
+#' @param with_post A [logical] value specifying whether a raw `jags.post`
+#' component
+#' should be included as an element of the [list] object returned by `run_mod()`
+#' (see `Value` section below for details).
+#' Note: These objects can be large.
+#' @returns A [dplyr::tbl_df] that contains the posterior
+#' distribution will be exported with the following variables:
+#'   - `iteration` = Number of sampling iterations
+#'   - `chain` = Number of MCMC chains run; between 1 and 4
+#'   - `Parameter` = Parameter being estimated. Includes the following:
+#'     - `y0` = Posterior estimate of baseline antibody concentration
+#'     - `y1` = Posterior estimate of peak antibody concentration
+#'     - `t1` = Posterior estimate of time to peak
+#'     - `shape` = Posterior estimate of shape parameter
+#'     - `alpha` = Posterior estimate of decay rate
+#'   - `Iso_type` = Antibody/antigen type combination being evaluated
+#'   - `Stratification` = The variable used to stratify jags model
+#'   - `Subject` = ID of subject being evaluated
+#'   - `value` = Estimated value of the parameter
+#' - The following [attributes] are included in the output:
+#'   - `class`: Class of the output object.
+#'   - `nChain`: Number of chains run.
+#'   - `nParameters`: The amount of parameters estimated in the model.
+#'   - `nIterations`: Number of iteration specified.
+#'   - `nBurnin`: Number of burn ins.
+#'   - `nThin`: Thinning number (niter/nmc)
+#'   - `priors`: A [list] that summarizes the input priors, including:
+#'     - `mu_hyp_param`
+#'     - `prec_hyp_param`
+#'     - `omega_param`
+#'     - `wishdf`
+#'     - `prec_logy_hyp_param`
+#'     - An optional `"jags.post"` attribute, included when argument
+#'   `with_post` = TRUE.
+#' @inheritDotParams prep_priors
 #' @export
 #' @example inst/examples/run_mod-examples.R
 run_mod <- function(data,
-                    file_mod,
+                    file_mod = serodynamics_example("model.jags"),
                     nchain = 4,
                     nadapt = 0,
                     nburn = 0,
                     nmc = 100,
                     niter = 100,
-                    strat = NA) {
+                    strat = NA,
+                    with_post = FALSE,
+                    ...) {
   ## Conditionally creating a stratification list to loop through
-  if (is.na(strat) == FALSE) {
-    strat_list <- unique(data[[strat]])
-  } else {
+  if (is.na(strat)) {
     strat_list <- "None"
+  } else {
+    strat_list <- unique(data[[strat]])
   }
 
   ## Creating a shell to output results
@@ -79,16 +93,17 @@ run_mod <- function(data,
   # For loop for running stratifications
   for (i in strat_list) {
     # Creating if else statement for running the loop
-    if (is.na(strat) == FALSE) {
+    if (is.na(strat)) {
+      dl_sub <- data
+    } else {
       dl_sub <- data |>
         dplyr::filter(.data[[strat]] == i)
-    } else {
-      dl_sub <- data
     }
 
     # prepare data for modeline
     longdata <- prep_data(dl_sub)
-    priors <- prep_priors(max_antigens = longdata$n_antigen_isos)
+    priorspec <- prep_priors(max_antigens = longdata$n_antigen_isos,
+                             ...)
 
     # inputs for jags model
     nchains <- nchain # nr of MC chains to run simultaneously
@@ -102,7 +117,7 @@ run_mod <- function(data,
 
     jags_post <- runjags::run.jags(
       model = file_mod,
-      data = c(longdata, priors),
+      data = c(longdata, priorspec),
       inits = initsfunction,
       method = "parallel",
       adapt = nadapt,
@@ -114,25 +129,23 @@ run_mod <- function(data,
       summarise = FALSE
     )
     # Assigning the raw jags output to a list.
-    # This will include a raw output for the jags.post for each stratification.
+    # This object will include a raw output for the jags.post for each
+    # stratification and will only be included if specified. 
     jags_post_final[[i]] <- jags_post
 
-    ## Cleaning the jags output -- much of this has to do with correctly
-    # classifying the [x,x] number
-    # included in the outputs
-    # ggs works with mcmc objects
+    # Unpacking and cleaning MCMC output.
     jags_unpack <- ggmcmc::ggs(jags_post[["mcmc"]])
 
     # Adding attributes
     mod_atts <- attributes(jags_unpack)
-    # Only keeping necesarry attributes
-    mod_atts <- mod_atts[3:8]
+    # Only keeping necessary attributes
+    mod_atts <- mod_atts[4:8]
 
     # extracting antigen-iso combinations to correctly number
     # then by the order they are estimated by the program.
     iso_dat <- data.frame(attributes(longdata)$antigens)
     iso_dat <- iso_dat |> dplyr::mutate(Subnum = as.numeric(row.names(iso_dat)))
-    ### Working with jags unpacked ggs outputs to clarify parameter and subject
+    # Working with jags unpacked ggs outputs to clarify parameter and subject
     jags_unpack <- jags_unpack |>
       dplyr::mutate(
         Subnum = sub(".*,", "", .data$Parameter),
@@ -145,14 +158,14 @@ run_mod <- function(data,
       )
     # Merging isodat in to ensure we are classifying antigen_iso
     jags_unpack <- dplyr::left_join(jags_unpack, iso_dat, by = "Subnum")
-    jags_unpack <- jags_unpack |>
-      dplyr::rename(c("Iso_type" = "attributes.longdata..antigens")) |>
-      dplyr::select(!c("Subnum"))
-    # Setting subset for the "new person" --setting it to the final sample
-    np <- as.character(longdata$nsubj)
+    ids <- data.frame(attr(longdata, "ids")) |>
+      mutate(Subject = as.character(dplyr::row_number()))
+    jags_unpack <- dplyr::left_join(jags_unpack, ids, by = "Subject")
     jags_final <- jags_unpack |>
-      dplyr::filter(.data$Subject == np)
-    ## Creating a label for the stratification, if there is one.
+      dplyr::select(!c("Subnum", "Subject")) |>
+      dplyr::rename(c("Iso_type" = "attributes.longdata..antigens",
+                      "Subject" = "attr.longdata...ids.."))
+    # Creating a label for the stratification, if there is one.
     # If not, will add in "None".
     jags_final$Stratification <- i
     ## Creating output
@@ -161,12 +174,26 @@ run_mod <- function(data,
   # Ensuring output does not have any NAs
   jags_out <- jags_out[complete.cases(jags_out), ]
   # Outputting the finalized jags output as a data frame with the
-  # jags output results for each stratification
-  # rbinded.
-  jags_out <- list(
-    "curve_params" = jags_out,
-    "jags.post" = jags_post_final,
-    "attributes" = mod_atts
-  )
+  # jags output results for each stratification rbinded.
+
+  # Making output a tibble and restructing.
+  jags_out <- dplyr::as_tibble(jags_out)  |>
+    select(!c("Parameter")) |>
+    rename("Parameter" = "Parameter_sub")
+  jags_out <- jags_out[, c("Iteration", "Chain", "Parameter", "Iso_type",
+                           "Stratification", "Subject", "value")]
+  current_atts <- attributes(jags_out) 
+  current_atts <- c(current_atts, mod_atts)
+  attributes(jags_out) <- current_atts
+  
+  # Adding priors
+  jags_out <- jags_out |>
+    structure("priors" = attributes(priorspec)$used_priors)
+  
+  # Conditionally adding jags.post
+  if (with_post) {
+    jags_out <- jags_out |>
+      structure(jags.post = jags_post_final)
+  } 
   jags_out
 }