+++ tests/testthat/test-run_mod.R
@@ -2,9 +2,7 @@ test_that(
   desc = "results are consistent with simulated data",
   code = {
     skip_on_os(c("windows", "linux"))
-    library(runjags)
     withr::local_seed(1)
-    library(dplyr)
     strat1 <- serocalculator::typhoid_curves_nostrat_100 |>
       sim_case_data(n = 100,
                     antigen_isos = "HlyE_IgA") |>
@@ -14,7 +12,7 @@ test_that(
       sim_case_data(n = 100,
                     antigen_isos = "HlyE_IgA") |>
       mutate(strat = "stratum 1")
-    dataset <- bind_rows(strat1, strat2)
+    dataset <- dplyr::bind_rows(strat1, strat2)
     withr::with_seed(
       1,
       code = {
@@ -27,10 +25,9 @@ test_that(
           nburn = 100, # Number of unrecorded samples before sampling begins
           nmc = 10,
           niter = 10, # Number of iterations
-          strat = "strat" # Variable to be stratified
+          strat = "strat", # Variable to be stratified
         ) |>
-          suppressWarnings() |>
-          magrittr::use_series("curve_params")
+          suppressWarnings()
         
         results |>
           attributes() |>
@@ -50,30 +47,19 @@ test_that(
   code = {
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
-    library(runjags)
-    dataset <- serodynamics_example(
-      "SEES_Case_Nepal_ForSeroKinetics_02-13-2025.csv"
-    ) |>
-      readr::read_csv() |>
-      as_case_data(
-        id_var = "person_id",
-        biomarker_var = "antigen_iso",
-        value_var = "result",
-        time_in_days = "dayssincefeveronset"
-      )
+    dataset <- serodynamics::nepal_sees 
 
     results <- run_mod(
       data = dataset, # The data set input
       file_mod = serodynamics_example("model.jags"),
       nchain = 2, # Number of mcmc chains to run
-      nadapt = 100, # Number of adaptations to run
-      nburn = 100, # Number of unrecorded samples before sampling begins
-      nmc = 10,
-      niter = 10, # Number of iterations
-      strat = "bldculres" # Variable to be stratified
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = "bldculres", # Variable to be stratified
     ) |>
-      suppressWarnings() |>
-      magrittr::use_series("curve_params")
+      suppressWarnings()
 
     results |>
       attributes() |>
@@ -90,34 +76,19 @@ test_that(
   code = {
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
-    library(runjags)
-    dataset <- serodynamics_example(
-      "SEES_Case_Nepal_ForSeroKinetics_02-13-2025.csv"
-    ) |>
-      readr::read_csv() |>
-      dplyr::mutate(
-        .by = person_id,
-        visit_num = dplyr::row_number()
-      ) |>
-      as_case_data(
-        id_var = "person_id",
-        biomarker_var = "antigen_iso",
-        value_var = "result",
-        time_in_days = "dayssincefeveronset"
-      )
+    dataset <- serodynamics::nepal_sees 
 
     results <- run_mod(
       data = dataset, # The data set input
       file_mod = serodynamics_example("model.jags"),
       nchain = 2, # Number of mcmc chains to run
-      nadapt = 100, # Number of adaptations to run
-      nburn = 100, # Number of unrecorded samples before sampling begins
-      nmc = 10,
-      niter = 10, # Number of iterations
-      strat = NA # Variable to be stratified
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = NA, # Variable to be stratified
     ) |>
-      suppressWarnings() |>
-      magrittr::use_series("curve_params")
+      suppressWarnings()
 
     results |>
       attributes() |>
@@ -126,5 +97,71 @@ test_that(
 
     results |>
       ssdtools:::expect_snapshot_data("nostrat-curve-params")
+  }
+)
+
+test_that(
+  desc = "results are consistent with unstratified SEES data with jags.post
+  included",
+  code = {
+    skip_on_os(c("windows", "linux"))
+    withr::local_seed(1)
+    dataset <- serodynamics::nepal_sees 
+    
+    results <- run_mod(
+      data = dataset, # The data set input
+      file_mod = serodynamics_example("model.jags"),
+      nchain = 2, # Number of mcmc chains to run
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = NA, # Variable to be stratified
+      with_post = TRUE
+    ) |>
+      suppressWarnings()
+    
+    results |>
+      attributes() |>
+      rlist::list.remove(c("row.names", "jags.post")) |>
+      expect_snapshot_value(style = "serialize")
+    
+    results |>
+      ssdtools:::expect_snapshot_data("nostrat-curve-params-withpost")
+  }
+)
+
+test_that(
+  desc = "results are consistent with unstratified SEES data with modified 
+  priors",
+  code = {
+    skip_on_os(c("windows", "linux"))
+    withr::local_seed(1)
+    dataset <- serodynamics::nepal_sees 
+    
+    results <- run_mod(
+      data = dataset, # The data set input
+      file_mod = serodynamics_example("model.jags"),
+      nchain = 2, # Number of mcmc chains to run
+      nadapt = 10, # Number of adaptations to run
+      nburn = 10, # Number of unrecorded samples before sampling begins
+      nmc = 100,
+      niter = 100, # Number of iterations
+      strat = NA, # Variable to be stratified
+      mu_hyp_param = c(1, 4, 1, -3, -1),
+      prec_hyp_param = c(0.01, 0.0001, 0.01, 0.001, 0.01),
+      omega_param = c(1, 20, 1, 10, 1),
+      wishdf_param = 10,
+      prec_logy_hyp_param = c(3, 1)
+    ) |>
+      suppressWarnings()
+    
+    results |>
+      attributes() |>
+      rlist::list.remove(c("row.names")) |>
+      expect_snapshot_value(style = "serialize")
+    
+    results |>
+      ssdtools:::expect_snapshot_data("nostrat-curve-params-specpriors")
   }
 )